feat(config): 更新文本渲染接口和配置项
- 在 README.md 中更新了 `/render/{user_id:str}` 接口的参数列表，增加了 `timeout` 参数
- 将 `WKHTMLTOPDF_PATH` 配置项重命名为 `WKHTMLTOIMAGE_PATH`，并更新了相应的路径
- 在 run_fastapi.py 中调整了 `render` 函数的 `timeout` 参数类型，从 int 改为 float

+++ run_fastapi.py
@@ -151,7 +151,7 @@ async def render(
     user_id: str,
     text: str = Form(...),
     style: str | None = Form(None),
-    timeout: int | None = Form(None),
+    timeout: float | None = Form(None),
 ):
     """
     Endpoint for rendering markdown text to image
@@ -166,7 +166,7 @@ async def render(
     rendered_image_dir = configs.get_config("rendered_image_dir", "./temp/render").get_value(Path)
 
     # 延迟删除函数
-    async def _wait_delete(sleep_time: int, filename: str):
+    async def _wait_delete(sleep_time: float, filename: str):
         """
         等待一段时间后删除图片
         """
